chore: expose twilioTokenValidator function, add twilio validator as dispatcher dependency
