+++ hrm-domain/lambdas/custom-integrations/uscr-dispatcher/index.ts
@@ -16,6 +16,7 @@
 
 // eslint-disable-next-line prettier/prettier
 import type { ALBEvent, ALBResult } from 'aws-lambda';
+import { twilioTokenValidator } from '@tech-matters/twilio-worker-auth';
 
 export const handler = async (event: ALBEvent): Promise<ALBResult> => {
   try {
