+++ packages/twilio-worker-auth/src/index.ts
@@ -14,6 +14,7 @@
  * along with this program.  If not, see https://www.gnu.org/licenses/.
  */
 
+export { twilioTokenValidator } from './twilioTokenValidator';
 export {
   getAuthorizationMiddleware,
   adminAuthorizationMiddleware,
